refactor and fix imports
