+++ src/index.tsx
@@ -1,5 +1,4 @@
 import { createRoot } from 'react-dom/client';
-
 import App from './App';
 
 const rootElement = document.getElementById('root');
